Tests/LibWeb/Text/expected/fullscreen/request-fullscreen-event-ordering.txt

resize 1920x1080
fullscreenchange
